+++ Session/Meta/AppDelegate.swift
@@ -40,20 +40,4 @@ extension AppDelegate {
     @objc func stopClosedGroupPoller() {
         ClosedGroupPoller.shared.stop()
     }
-    
-    @objc func getAppModeOrSystemDefault() -> AppMode {
-        let userDefaults = UserDefaults.standard
-        
-        guard userDefaults.dictionaryRepresentation().keys.contains("appMode") else {
-            if #available(iOS 13.0, *) {
-                return UITraitCollection.current.userInterfaceStyle == .dark ? .dark : .light
-            } else {
-                return .light
-            }
-        }
-        
-        let mode = userDefaults.integer(forKey: "appMode")
-        return AppMode(rawValue: mode) ?? .light
-    }
-    
 }
